feat: add layout instead of ApproachesWrapper
