+++ app/globals.css
@@ -107,8 +107,8 @@ li {
 
   @media (max-width: 820px) {
     position: fixed;
-    top: calc(50vh - 30px);
-    left: calc(50vw - 30px);
+    top: calc(50vh - 15px);
+    left: calc(50vw - 15px);
     width: 30px;
     height: 30px;
   }
